rename xmovie -> xanimations
