+++ xanimations/test/test_presets.py
@@ -5,7 +5,7 @@ import numpy as np
 import pytest
 import xarray as xr
 
-from xmovie.presets import (
+from xanimations.presets import (
     _check_input,
     _core_plot,
     _smooth_boundary_NearsidePerspective,
